Add link action allowing linking files within image

+++ lib/imagemaster3000/utils/tmp.rb
@@ -1,4 +1,3 @@
-require 'open-uri'
 require 'tempfile'
 
 module Imagemaster3000
@@ -6,8 +5,8 @@ module Imagemaster3000
     class Tmp
       def self.download(url)
         file = Tempfile.new('imagemaster3000')
-        download = open(url)
         logger.debug "Downloading file from url #{url} to tempfile #{file.path.inspect}"
+        download = URI.parse(url).open
         IO.copy_stream(download, file)
 
         file.rewind
@@ -15,6 +14,8 @@ module Imagemaster3000
       end
 
       def self.destroy(file)
+        return unless file
+
         logger.debug "Closing tempfile #{file.path.inspect}"
         file.close
         file.unlink
